requirements.txt
Flask
Flask-WTF
WTForms
psycopg2-binary
python-dotenv